Fix issue for building on arm64/linux
- fixes https://github.com/VKCOM/engine-go/issues/1
